Rename withAuthUser file
